refactoring of all the folders and files for better codding and understanding
