+++ src/components/common/loadingSpinner.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import Loader from "react-loader-spinner";
 import "react-loader-spinner/dist/loader/css/react-spinner-loader.css";
+
 export default function LoadingSpinner() {
   return (
     <div>
