+++ src/components/styles/posts.css
@@ -1,0 +1,171 @@
+/* post details styles  */
+
+.post-details-wrapper {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+}
+.post-detail-header {
+  width: 100%;
+  height: 200px;
+  background-color: #fc427b;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.post-category {
+  padding: 3px 10px;
+  border-radius: 5px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: var(--white);
+  margin-top: 60px;
+}
+.post-category span {
+  color: #fc427b;
+  font-weight: 600;
+  background: whatever;
+}
+.question-synthese {
+  font-weight: 500;
+  color: var(--white);
+  font-size: 20px;
+  line-height: 0;
+}
+.post-details-container {
+  width: 80%;
+  min-height: 700px;
+  position: relative;
+  /* padding-right: 200px; */
+}
+
+.post-detail-comment {
+  position: fixed;
+  top: 200px;
+  width: 200px;
+  height: 100%;
+  right: 0;
+}
+.post-sidebar-wrapper {
+  padding: 30px 0;
+}
+.add-comment-btn {
+  width: 150px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 40px;
+  border-radius: 5px;
+  background-color: var(--primary-color);
+  cursor: pointer;
+}
+.add-comment-btn span {
+  color: var(--white);
+  font-weight: 600;
+  font-weight: 14px;
+}
+.post-detail-profile {
+  padding-top: 30px;
+  display: flex;
+  padding-right: 100px;
+  flex-direction: column;
+}
+.post-profile {
+  display: flex;
+  align-items: center;
+}
+.post-owner {
+  width: 60px;
+  height: 60px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: #fc427b;
+  border-radius: 50%;
+  margin-right: 20px;
+  cursor: pointer;
+  text-decoration: none;
+  overflow: hidden;
+}
+.post-owner img {
+  width: 100%;
+  height: 100%;
+}
+.post-owner span {
+  font-size: 25px;
+  color: var(--white);
+}
+.post-timing-details {
+  display: flex;
+  flex-direction: column;
+  text-align: left;
+  padding: 10px 0;
+  text-decoration: none;
+  align-items: center !important;
+}
+.owner-details {
+  line-height: 0;
+  text-align: left;
+  color: #525050;
+  text-decoration: none;
+  margin-bottom: 15px;
+}
+
+.timing {
+  color: var(--light);
+}
+.timing span:nth-child(1) {
+  margin-right: 15px;
+}
+.post-container {
+  margin-right: 100px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding-top: 0px;
+  padding-left: 80px;
+  text-align: left;
+  border-bottom: 1px solid var(--border-bottom-color);
+  padding-bottom: 20px;
+}
+.post-container .content p {
+  line-height: 1.7;
+}
+.footer {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  width: 100%;
+}
+.footer .post-btn {
+  padding-right: 20px;
+  display: flex;
+  /* flex-direction: column; */
+  align-items: center;
+}
+.footer .post-btn span {
+  margin: 0 15px;
+  cursor: pointer;
+}
+.additional-info-post {
+  display: flex;
+  align-items: center;
+  margin-left: 10px;
+  justify-content: center;
+  cursor: pointer;
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  transition: all 0.4s;
+  position: relative;
+}
+.additional-info-post:hover {
+  background-color: var(--primary-color);
+  color: var(--white);
+}
+/* end post details  */
