+++ src/components/textEditor/customMentionPlugin.jsx
@@ -1,9 +1,9 @@
 import React, { Component } from "react";
 import { EditorState } from "draft-js";
-import Editor from "draft-js-plugins-editor";
 import createMentionPlugin, {
   defaultSuggestionsFilter,
 } from "draft-js-mention-plugin";
+
 import editorStyles from "./editorStyles.css";
 import mentionsStyles from "./mentionsStyles.css";
 import mentions from "./mentions";
@@ -98,18 +98,9 @@ export default class CustomMentionEditor extends Component {
 
   render() {
     const { MentionSuggestions } = this.mentionPlugin;
-    // const plugins = [this.mentionPlugin];
 
     return (
       <div className={editorStyles.editor} onClick={this.focus}>
-        {/* <Editor
-          editorState={this.state.editorState}
-          onChange={this.onChange}
-          plugins={plugins}
-          ref={(element) => {
-            this.editor = element;
-          }}
-        /> */}
         <MentionSuggestions
           onSearchChange={this.onSearchChange}
           suggestions={this.state.suggestions}
