+++ src/components/textEditor/suggestionEntry.jsx
@@ -1,5 +1,7 @@
 import React from "react";
+
 import "./mentionsStyles.css";
+
 const Entry = (props) => {
   const {
     mention,
